Creación de endpoints en el controlador Producto y diseño funcionalidad en las vistas de la entidad producto

+++ src/main/java/com/sistema/ventas/vitapan/entities/Cliente.java
@@ -20,5 +20,5 @@ public class Cliente {
     private String apellido;
     private String direccion;
     private String telefono;
-    private byte estado;
+    private int estado;
 }
